requirements.txt
requests
ratelimit
Pillow
ImageHash
python-dotenv
psycopg2-binary